code updated

+++ Domain/Entities/CoinModel.cs
@@ -11,8 +11,8 @@ namespace Domain.Entities
         public long Id { get; set; }
         public long UserId { get; set; }
         public string UserName { get; set; }
-        public double Coin { get; set; }
-        public double CoinType { get; set; }//cr-c/dr-d
+        public int Coin { get; set; }
+        public int CoinType { get; set; }//cr-c/dr-d
         public string CoinColor { get; set; }//red//green
     }
 }
